requirements.txt
# Core dependencies for Dakota language extraction
# This project uses Claude Sonnet 4.5 via Anthropic API - no local model inference needed

# Anthropic API for Claude Sonnet 4.5 (primary VLM)
anthropic>=0.39.0

# Google Gemini API for Q&A generation
google-generativeai>=0.3.0

# HTTP requests for API calls
requests>=2.31.0

# Image processing for vision tasks
Pillow>=10.0.0

# Dataset publishing and model hub integration
huggingface-hub>=0.24.0

# Environment variable management
python-dotenv>=1.0.0

# Progress bars and utilities
tqdm>=4.66.0

# Weights & Biases for experiment tracking and logging
# Weights & Biases for experiment tracking and logging
wandb>=0.16.0
weave>=0.50.0

# Thinking Machines / Tinker integration
tinker>=0.4.0
chz>=0.1.1
git+https://github.com/thinking-machines-lab/tinker-cookbook.git
